pyproject.toml
[project]
name = "stock-agent"
version = "0.1.0"
description = "AI-powered stock market analysis agent with real-time data, SEC filings RAG, and LangGraph orchestration"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "chainlit>=2.11.0",
    "chromadb>=1.5.8",
    "edgar>=5.6.3",
    "edgartools>=5.30.0",
    "eventregistry>=9.1",
    "faiss-cpu>=1.13.2",
    "langchain>=1.2.15",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.1.0",
    "langchain-openai>=1.1.13",
    "langchain-pinecone>=0.2.13",
    "langgraph>=1.1.6",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "lxml>=6.0.4",
    "numexpr>=2.14.1",
    "pinecone-client>=6.0.0",
    "polygon>=1.2.8",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "rich>=15.0.0",
    "tiktoken>=0.12.0",
    "yfinance>=1.2.2",
]
